requirements.txt
folium==0.14.0
geopandas==0.12.2
Flask==2.2.3
gunicorn==20.1.0